refinining workers. Getting ready to migrate to more non-blocking logic
