+++ Unosquare.FFME.Common/Commands/SeekCommand.cs
@@ -2,7 +2,6 @@
 {
     using Shared;
     using System;
-    using System.Threading;
 
     /// <summary>
     /// The Seek Command Implementation
@@ -97,8 +96,7 @@
                 // to finish. We don't want to interfere with reading in progress
                 // or decoding in progress. For decoding we already know we are not
                 // in a cycle because the decoding worker called this logic.
-                m.Workers.ReadingWorker.PauseAsync();
-                m.Workers.ReadingWorker.Wait(Timeout.Infinite);
+                m.Workers.ReadingWorker.PauseAsync().Wait();
 
                 // Signal the starting state clearing the packet buffer cache
                 m.Container.Components.ClearQueuedPackets(flushBuffers: true);
