+++ Unosquare.FFME.Common/Primitives/WorkerBase.cs
@@ -6,47 +6,30 @@
     using System.Runtime.CompilerServices;
     using System.Threading;
     using System.Threading.Tasks;
-    using ThreadState = System.Threading.ThreadState;
 
     /// <summary>
-    /// Provides a base implementation for application workers
-    /// that perform continuous, long-running tasks. This class
-    /// provides the ability to perform fine-grained control on these tasks.
+    /// Provides base infrastructure for Timer and Thread workers.
     /// </summary>
     /// <seealso cref="IWorker" />
     public abstract class WorkerBase : IWorker
     {
-        private readonly object SyncLock = new object();
-        private readonly Dictionary<StateChangeRequest, bool> StateChangeRequests;
-        private readonly ManualResetEventSlim CycleCompletedEvent = new ManualResetEventSlim(true);
-        private readonly ManualResetEventSlim StateChangedEvent = new ManualResetEventSlim(true);
-
-        private readonly Thread Thread;
-        private readonly Stopwatch CycleStopwatch = new Stopwatch();
-
         // Since these are API property backers, we use interlocked to read from them
         // to avoid deadlocked reads
+        private readonly object SyncLock = new object();
         private long m_Period;
         private int m_IsDisposed = 0;
         private int m_IsDisposing = 0;
         private int m_WorkerState = (int)WorkerState.Created;
-        private Task<WorkerState> StateChangeTask;
-
-        // This will be recreated on demand
-        private CancellationTokenOwner CycleCancellation = new CancellationTokenOwner();
 
         /// <summary>
         /// Initializes a new instance of the <see cref="WorkerBase"/> class.
         /// </summary>
         /// <param name="name">The name.</param>
-        /// <param name="priority">The thread priority</param>
-        /// <param name="period">The interval of cycle execution</param>
-        /// <param name="delayProvider">The cycle delay provide implementation</param>
-        protected WorkerBase(string name, ThreadPriority priority, TimeSpan period, IWorkerDelayProvider delayProvider)
+        /// <param name="period">The execution interval.</param>
+        protected WorkerBase(string name, TimeSpan period)
         {
-            Period = period;
             Name = name;
-            DelayProvider = delayProvider;
+            Period = period;
 
             StateChangeRequests = new Dictionary<StateChangeRequest, bool>(5)
             {
@@ -55,19 +38,12 @@
                 [StateChangeRequest.Resume] = false,
                 [StateChangeRequest.Stop] = false
             };
-
-            Thread = new Thread(RunWorkerLoop)
-            {
-                IsBackground = true,
-                Priority = priority,
-                Name = name
-            };
         }
 
         /// <summary>
         /// Enumerates all the different state change requests
         /// </summary>
-        private enum StateChangeRequest
+        protected enum StateChangeRequest
         {
             /// <summary>
             /// No state change request.
@@ -109,311 +85,73 @@
         public WorkerState WorkerState
         {
             get => (WorkerState)Interlocked.CompareExchange(ref m_WorkerState, 0, 0);
-            private set => Interlocked.Exchange(ref m_WorkerState, (int)value);
+            protected set => Interlocked.Exchange(ref m_WorkerState, (int)value);
         }
 
         /// <inheritdoc />
         public bool IsDisposed
         {
             get => Interlocked.CompareExchange(ref m_IsDisposed, 0, 0) != 0;
-            private set => Interlocked.Exchange(ref m_IsDisposed, value ? 1 : 0);
+            protected set => Interlocked.Exchange(ref m_IsDisposed, value ? 1 : 0);
         }
 
         /// <inheritdoc />
         public bool IsDisposing
         {
             get => Interlocked.CompareExchange(ref m_IsDisposing, 0, 0) != 0;
-            private set => Interlocked.Exchange(ref m_IsDisposing, value ? 1 : 0);
+            protected set => Interlocked.Exchange(ref m_IsDisposing, value ? 1 : 0);
         }
 
         /// <summary>
-        /// Provides an implementation on a cycle delay provider.
+        /// Gets the cycle stopwatch.
         /// </summary>
-        protected IWorkerDelayProvider DelayProvider { get; }
-
-        /// <inheritdoc />
-        public Task<WorkerState> StartAsync()
-        {
-            lock (SyncLock)
-            {
-                if (WorkerState == WorkerState.Paused || WorkerState == WorkerState.Waiting)
-                    return ResumeAsync();
-
-                if (WorkerState != WorkerState.Created)
-                    return Task.FromResult(WorkerState);
-
-                var task = QueueStateChange(StateChangeRequest.Start);
-                Thread.Start();
-                return task;
-            }
-        }
-
-        /// <inheritdoc />
-        public Task<WorkerState> PauseAsync()
-        {
-            lock (SyncLock)
-            {
-                if (WorkerState != WorkerState.Running && WorkerState != WorkerState.Waiting)
-                    return Task.FromResult(WorkerState);
-
-                var task = QueueStateChange(StateChangeRequest.Pause);
-                return task;
-            }
-        }
-
-        /// <inheritdoc />
-        public Task<WorkerState> ResumeAsync()
-        {
-            lock (SyncLock)
-            {
-                if (WorkerState == WorkerState.Created)
-                    return StartAsync();
-
-                if (WorkerState != WorkerState.Paused && WorkerState != WorkerState.Waiting)
-                    return Task.FromResult(WorkerState);
-
-                var task = QueueStateChange(StateChangeRequest.Resume);
-                return task;
-            }
-        }
-
-        /// <inheritdoc />
-        public Task<WorkerState> StopAsync()
-        {
-            lock (SyncLock)
-            {
-                if (WorkerState == WorkerState.Stopped || WorkerState == WorkerState.Created)
-                {
-                    WorkerState = WorkerState.Stopped;
-                    return Task.FromResult(WorkerState);
-                }
-
-                var task = QueueStateChange(StateChangeRequest.Stop);
-                return task;
-            }
-        }
+        protected Stopwatch CycleStopwatch { get; } = new Stopwatch();
 
         /// <summary>
-        /// Waits for cycle.
+        /// Gets the state change requests.
         /// </summary>
-        /// <param name="millisecondsTimeout">The milliseconds timeout.</param>
-        /// <returns>True if the wait was successful</returns>
-        public bool Wait(int millisecondsTimeout)
-        {
-            if (IsDisposing || IsDisposed) return false;
-            return CycleCompletedEvent.Wait(millisecondsTimeout);
-        }
+        protected Dictionary<StateChangeRequest, bool> StateChangeRequests { get; }
+
+        /// <summary>
+        /// Gets the cycle completed event.
+        /// </summary>
+        protected ManualResetEventSlim CycleCompletedEvent { get; } = new ManualResetEventSlim(true);
+
+        /// <summary>
+        /// Gets the state changed event.
+        /// </summary>
+        protected ManualResetEventSlim StateChangedEvent { get; } = new ManualResetEventSlim(true);
+
+        /// <summary>
+        /// Gets the cycle logic cancellation owner.
+        /// </summary>
+        protected CancellationTokenOwner CycleCancellation { get; } = new CancellationTokenOwner();
+
+        /// <summary>
+        /// Gets or sets the state change task.
+        /// </summary>
+        protected Task<WorkerState> StateChangeTask { get; set; }
+
+        /// <inheritdoc />
+        public abstract Task<WorkerState> StartAsync();
+
+        /// <inheritdoc />
+        public abstract Task<WorkerState> PauseAsync();
+
+        /// <inheritdoc />
+        public abstract Task<WorkerState> ResumeAsync();
+
+        /// <inheritdoc />
+        public abstract Task<WorkerState> StopAsync();
 
         /// <inheritdoc />
         public void Dispose() => Dispose(true);
 
         /// <summary>
-        /// Suspends execution queues a new new cycle for execution. The delay is given in
-        /// milliseconds. When overridden in a derived class the wait handle will be set
-        /// whenever an interrupt is received.
-        /// </summary>
-        /// <param name="wantedDelay">The remaining delay to wait for in the cycle</param>
-        /// <param name="delayTask">Contains a reference to a task with the scheduled period delay</param>
-        /// <param name="token">The cancellation token to cancel waiting</param>
-        protected virtual void ExecuteCycleDelay(int wantedDelay, Task delayTask, CancellationToken token) =>
-            DelayProvider?.ExecuteCycleDelay(wantedDelay, delayTask, token);
-
-        /// <summary>
-        /// Represents the user defined logic to be executed on a single worker cycle.
-        /// Check the cancellation token continuously if you need responsive interrupts.
-        /// </summary>
-        /// <param name="ct">The cancellation token.</param>
-        protected abstract void ExecuteCycleLogic(CancellationToken ct);
-
-        /// <summary>
-        /// Handles the cycle logic exceptions.
-        /// </summary>
-        /// <param name="ex">The exception that was thrown.</param>
-        protected abstract void HandleCycleLogicException(Exception ex);
-
-        /// <summary>
-        /// This method is called automatically when <see cref="Dispose()"/> is called.
-        /// Makes sure you release all resources within this call.
-        /// </summary>
-        protected abstract void DisposeManagedState();
-
-        /// <summary>
-        /// Implements worker control, execution and delay logic in a loop.
-        /// </summary>
-        private void RunWorkerLoop()
-        {
-            while (WorkerState != WorkerState.Stopped)
-            {
-                CycleStopwatch.Restart();
-                var interruptToken = CycleCancellation.Token;
-                var period = Period.TotalMilliseconds >= int.MaxValue ? -1 : Convert.ToInt32(Math.Floor(Period.TotalMilliseconds));
-                var delayTask = Task.Delay(period, interruptToken);
-                var initialWorkerState = WorkerState;
-
-                // Lock the cycle and capture relevant state valid for this cycle
-                CycleCompletedEvent.Reset();
-
-                // Process the tasks that are awaiting
-                if (ProcessStateChangeRequests())
-                    continue;
-
-                try
-                {
-                    if (initialWorkerState == WorkerState.Waiting &&
-                        !interruptToken.IsCancellationRequested)
-                    {
-                        // Mark the state as Running
-                        WorkerState = WorkerState.Running;
-
-                        // Call the execution logic
-                        ExecuteCycleLogic(interruptToken);
-                    }
-                }
-                catch (Exception ex)
-                {
-                    HandleCycleLogicException(ex);
-                }
-                finally
-                {
-                    // Update the state
-                    WorkerState = initialWorkerState == WorkerState.Paused
-                        ? WorkerState.Paused
-                        : WorkerState.Waiting;
-
-                    // Signal the cycle has been completed so new cycles can be executed
-                    CycleCompletedEvent.Set();
-
-                    if (!interruptToken.IsCancellationRequested)
-                    {
-                        ExecuteCycleDelay(
-                            ComputeCycleDelay(initialWorkerState),
-                            delayTask,
-                            CycleCancellation.Token);
-                    }
-                }
-            }
-
-            ClearStateChangeRequests();
-            WorkerState = WorkerState.Stopped;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private int ComputeCycleDelay(WorkerState initialWorkerState)
-        {
-            var delay = 0;
-            var elapsedMillis = CycleStopwatch.ElapsedMilliseconds;
-            var period = Period;
-            var periodMillis = period.TotalMilliseconds;
-            var delayMillis = periodMillis - elapsedMillis;
-
-            if (initialWorkerState == WorkerState.Paused || period == TimeSpan.MaxValue || delayMillis >= int.MaxValue)
-                delay = Timeout.Infinite;
-            else if (elapsedMillis >= periodMillis)
-                delay = 0;
-            else
-                delay = Convert.ToInt32(Math.Floor(delayMillis));
-
-            return delay;
-        }
-
-        /// <summary>
-        /// Queues a transition in worker state for processing. Returns a task that can be awaited
-        /// when the operation completes.
-        /// </summary>
-        /// <param name="request">The request.</param>
-        /// <returns>The awaitable task.</returns>
-        private Task<WorkerState> QueueStateChange(StateChangeRequest request)
-        {
-            lock (SyncLock)
-            {
-                if (StateChangeTask != null)
-                    return StateChangeTask;
-
-                var waitingTask = new Task<WorkerState>(() =>
-                {
-                    StateChangedEvent.Wait();
-                    lock (SyncLock)
-                    {
-                        StateChangeTask = null;
-                        return WorkerState;
-                    }
-                });
-
-                waitingTask.ConfigureAwait(false);
-                StateChangeTask = waitingTask;
-                StateChangedEvent.Reset();
-                StateChangeRequests[request] = true;
-                waitingTask.Start();
-                CycleCancellation.Cancel();
-
-                return waitingTask;
-            }
-        }
-
-        /// <summary>
-        /// Processes the state change request by checking pending events and scheduling
-        /// cycle execution accordingly. The <see cref="WorkerState"/> is also updated.
-        /// </summary>
-        /// <returns>Returns <c>true</c> if the execution should be terminated. <c>false</c> otherwise.</returns>
-        private bool ProcessStateChangeRequests()
-        {
-            lock (SyncLock)
-            {
-                var hasRequest = false;
-
-                if (StateChangeRequests[StateChangeRequest.Start])
-                {
-                    hasRequest = true;
-                    WorkerState = WorkerState.Waiting;
-                }
-                else if (StateChangeRequests[StateChangeRequest.Pause])
-                {
-                    hasRequest = true;
-                    WorkerState = WorkerState.Paused;
-                }
-                else if (StateChangeRequests[StateChangeRequest.Resume])
-                {
-                    hasRequest = true;
-                    WorkerState = WorkerState.Waiting;
-                }
-                else if (StateChangeRequests[StateChangeRequest.Stop])
-                {
-                    hasRequest = true;
-                    WorkerState = WorkerState.Stopped;
-                }
-
-                // Signals all state changes to continue
-                // as a command has been handled.
-                if (hasRequest)
-                    ClearStateChangeRequests();
-
-                return hasRequest;
-            }
-        }
-
-        /// <summary>
-        /// Signals all state change requests to set.
-        /// </summary>
-        private void ClearStateChangeRequests()
-        {
-            lock (SyncLock)
-            {
-                // Mark all events as completed
-                StateChangeRequests[StateChangeRequest.Start] = false;
-                StateChangeRequests[StateChangeRequest.Pause] = false;
-                StateChangeRequests[StateChangeRequest.Resume] = false;
-                StateChangeRequests[StateChangeRequest.Stop] = false;
-
-                StateChangedEvent.Set();
-                CycleCompletedEvent.Set();
-            }
-        }
-
-        /// <summary>
         /// Releases unmanaged and - optionally - managed resources.
         /// </summary>
         /// <param name="alsoManaged"><c>true</c> to release both managed and unmanaged resources; <c>false</c> to release only unmanaged resources.</param>
-        private void Dispose(bool alsoManaged)
+        protected virtual void Dispose(bool alsoManaged)
         {
             lock (SyncLock)
             {
@@ -431,13 +169,51 @@
             CycleCompletedEvent.Dispose();
             CycleCancellation.Dispose();
 
-            if ((Thread.ThreadState & ThreadState.Unstarted) != ThreadState.Unstarted)
-                Thread.Join();
-
-            DisposeManagedState();
+            OnDisposing();
             IsDisposed = true;
+        }
 
-            // There are no unmanaged resources
+        /// <summary>
+        /// Handles the cycle logic exceptions.
+        /// </summary>
+        /// <param name="ex">The exception that was thrown.</param>
+        protected abstract void OnCycleException(Exception ex);
+
+        /// <summary>
+        /// Represents the user defined logic to be executed on a single worker cycle.
+        /// Check the cancellation token continuously if you need responsive interrupts.
+        /// </summary>
+        /// <param name="ct">The cancellation token.</param>
+        protected abstract void ExecuteCycleLogic(CancellationToken ct);
+
+        /// <summary>
+        /// This method is called automatically when <see cref="Dispose()"/> is called.
+        /// Makes sure you release all resources within this call.
+        /// </summary>
+        protected abstract void OnDisposing();
+
+        /// <summary>
+        /// Computes the cycle delay.
+        /// </summary>
+        /// <param name="initialWorkerState">Initial state of the worker.</param>
+        /// <returns>The number of milliseconds to delay for</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        protected int ComputeCycleDelay(WorkerState initialWorkerState)
+        {
+            var delay = 0;
+            var elapsedMillis = CycleStopwatch.ElapsedMilliseconds;
+            var period = Period;
+            var periodMillis = period.TotalMilliseconds;
+            var delayMillis = periodMillis - elapsedMillis;
+
+            if (initialWorkerState == WorkerState.Paused || period == TimeSpan.MaxValue || delayMillis >= int.MaxValue)
+                delay = Timeout.Infinite;
+            else if (elapsedMillis >= periodMillis)
+                delay = 0;
+            else
+                delay = Convert.ToInt32(Math.Floor(delayMillis));
+
+            return delay;
         }
     }
 }
